refactor: Move useAuth hook to dedicated file for Fast Refresh compliance
- Extract useAuth hook to features/auth/hooks/useAuth.ts
- Update AuthContext to remove useAuth export (Fast Refresh compliance)
- Update hooks/use-auth.ts barrel to export from new file
- Removes unused useContext import from AuthContext

All quality gates pass: lint, typecheck, tests, docs:generate
