+++ apps/customer-dashboard/src/features/auth/hooks/use-auth.ts
@@ -1,2 +1,2 @@
-export { useAuth } from '../context/AuthContext';
+export { useAuth } from './useAuth';
 export { AuthProvider } from '../context/AuthContext';
